feat(platforms): updating grants stamp description
